Add annotation dataclasses to pkg_types.py Fixes #106
