+++ pkg_api/nl_to_pkg/annotators/annotator.py
@@ -5,12 +5,11 @@ annotate the triple (subject, predicate, object) and the preference (1 |
 -1) in the query.
 """
 
-
 from abc import ABC, abstractmethod
 from typing import Tuple
 
-from pkg_api.core.annotation import PKGData
 from pkg_api.core.intents import Intent
+from pkg_api.core.pkg_types import PKGData
 from pkg_api.nl_to_pkg.llm.prompt import Prompt
