+++ pkg_api/nl_to_pkg/annotators/three_step_annotator.py
@@ -4,13 +4,12 @@ This module contains a three-step annotator for annotating a statement
 with a triple and a preference using LLM.
 """
 
-
 import re
 import uuid
 from typing import Dict, Optional, Tuple
 
-from pkg_api.core.annotation import PKGData, Preference, Triple, TripleElement
 from pkg_api.core.intents import Intent
+from pkg_api.core.pkg_types import PKGData, Preference, Triple, TripleElement
 from pkg_api.nl_to_pkg.annotators.annotator import StatementAnnotator
 from pkg_api.nl_to_pkg.llm.llm_connector import LLMConnector
 from pkg_api.nl_to_pkg.llm.prompt import Prompt
