+++ tests/nl_to_pkg/test_eval_nl_to_pkg.py
@@ -1,12 +1,13 @@
 """Tests eval_nl_to_pkg.py file."""
+
 import uuid
 from typing import Dict, List, Tuple
 from unittest.mock import MagicMock, patch
 
 import pytest
 
-from pkg_api.core.annotation import PKGData, Triple, TripleElement
 from pkg_api.core.intents import Intent
+from pkg_api.core.pkg_types import PKGData, Triple, TripleElement
 from pkg_api.nl_to_pkg.annotators.three_step_annotator import (
     ThreeStepStatementAnnotator,
 )
