+++ tests/nl_to_pkg/test_spotlight_entity_linker.py
@@ -1,11 +1,11 @@
 """Tests the Spotlight entity linker class."""
+
 import uuid
 from unittest.mock import Mock, patch
 
 import pytest
 
-from pkg_api.core.annotation import Concept, PKGData, Triple, TripleElement
-from pkg_api.core.pkg_types import URI
+from pkg_api.core.pkg_types import URI, Concept, PKGData, Triple, TripleElement
 from pkg_api.nl_to_pkg.entity_linking.spotlight_entity_linker import (
     SpotlightEntityLinker,
 )
